+++ CHANGELOG.md
@@ -5,7 +5,20 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.2.30 - Unreleased
+## Version 0.2.31 - Unreleased
+
+
+### Changed
+* kwcoco subset can now accept dst as the second positional argument
+
+
+### Fixed
+* Fix issue in kwcoco subset where paths were invalid if written outside of
+  the source kwcoco bundle.
+* Remove the gdal cache, which caused issues in read/write cases.
+
+
+## Version 0.2.30 - Released 2022-04-19
 
 ### Added
 * kwcoco validate now accepts `--require_relative` which causes non-portable